refactor(control): extract autoResearchManager collaborator
The AutoResearch subsystem (~320 lines: task prepare/resume, heartbeats,
evidence recording, direction summaries, readiness) lived inline on
Controller. Per the leaf-collaborator playbook it moves to
autoresearch_manager.go: autoResearchManager wraps the optional
workspace autoresearch.Store and never touches Controller state —
assistant text and task IDs are passed in. Controller keeps thin glue
that resolves the active task from the goal machine and owns notices
(prepareAutoResearchTask fast-path, finalizeAutoResearchTask, the
AutoResearch* port methods). controller.go 6498 -> 6179 lines.
